pyproject.toml
[project]
name = "hgc-helper"
version = "0.1.0"
description = "Educational Insights Platform for History, Geography, and Civics"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "llama-index>=0.10.12",
    "llama-index-core>=0.10.12",
    "llama-index-llms-openai>=0.1.5",
    "llama-index-llms-gemini>=0.1.4",
    "llama-index-embeddings-openai>=0.1.5",
    "llama-index-vector-stores-chroma>=0.1.4",
    "llama-index-retrievers-bm25>=0.1.3",
    "chromadb>=0.4.22",
    "openai>=1.12.0",
    "cohere>=4.47",
    "streamlit>=1.31.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.2.0",
    "numpy>=1.26.3",
    "nltk>=3.8.1",
    "plotly>=5.18.0",
    "wordcloud>=1.9.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
